requirements.txt
flask
requests
python-dotenv
openai==0.28
